Add files via upload
Assignment details:
Develop a program that uses a HashMap to output a histogram (bar chart) of numbers entered by the user. The number should keep asking the user to enter a number between 1 and 9 (inclusive) and stop when the user enters -1.  Then, it should draw a simple histogram showing how many times each digit was entered. For example, the following input: